pyproject.toml
[tool.poetry]
name = "automate-ci-cd"
version = "0.1.0"
description = ""
authors = ["bion-slmn <bionsol25@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.8"
django = "^4.2"
djangorestframework = "^3.15.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
